# src/api/swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_auditlog: Read audit log
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
      restart_api: Restart the API components
      access_tokens: Administer access tokens
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/apis':
    get:
      security:
      - accessCode: ['read_apis']
      - implicit: ['read_apis']
      - password: ['read_apis']
      - clientCredentials: ['read_apis']
      tags:
        - apis
      description: Gets the list of APIs of this API Portal.
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              apis:
                title: ArrayOfApis
                type: array
                items:
                  $ref: '#/definitions/apiType'
              _links:
                type: object
                properties:
                  self:
                    $ref: '#/definitions/hrefType'
                  plans:
                    $ref: '#/definitions/hrefType'
...
  '/apis/{apiId}/plans':
    get:
      security:
      - accessCode: ['read_apis', 'read_plans']
      - implicit: ['read_apis', 'read_plans']
      - password: ['read_apis', 'read_plans']
      - clientCredentials: ['read_apis', 'read_plans']
      tags:
        - apis
      description: Returns a description of the API in markdown notation.
      produces:
        - text/markdown
      parameters:
        - name: apiId
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/planType'
        '403':
          description: Not allowed
        '404':
          description: Not found
...
  '/apis/{apiId}/subscriptions':
    get:
      security:
      - accessCode: ['read_apis', 'read_subscriptions']
      - implicit: ['read_apis', 'read_subscriptions']
      - password: ['read_apis', 'read_subscriptions']
      tags:
        - apis
      description: Retrieve all applications subscribed to this API
      produces:
        - application/json
      parameters:
        - name: apiId
          in: path
          type: string
          required: true
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              count:
                type: number
              count_cached:
                type: boolean
              offset:
                type: number
              limit:
                type: number
              items:
                type: array
                items:
                  type: object
                  properties:
                    application:
                      type: string
                    plan:
                      type: string
        403:
          description: Not allowed (Admins only)
        404:
          description: Not found (unknown API)
  '/plans':
    get:
      security:
      - accessCode: ['read_plans']
      - implicit: ['read_plans']
      - password: ['read_plans']
      - clientCredentials: ['read_plans']
      tags:
        - misc
      description: Gets the available plans for APIs
      responses:
        '200':
          description: Successful response
          schema:
            title: ArrayOfPlans
            type: array
            items:
              $ref: '#/definitions/planType'
  '/groups':
    get:
      security:
      - accessCode: ['read_groups']
      - implicit: ['read_groups']
      - password: ['read_groups']
      - clientCredentials: ['read_groups']
      tags:
        - misc
      description: Gets the list of available user groups
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              groups:
                type: array
                items:
                  title: Group
                  type: object
                  properties:
                    id:
                      type: string
                    alt_ids:
                      type: array
                      items:
                        type: string
                    name:
                      type: string
                    adminGroup:
                      type: boolean
              _links:
                type: object
                properties:
                  self:
                    $ref: '#/definitions/hrefType'
  '/users':
    get:
      security:
      - accessCode: ['read_users']
      - implicit: ['read_users']
      - password: ['read_users']
      tags:
        - users
      parameters:
        - name: customId
          description: Pass to retrieve user by customId
          in: query
          type: string
          required: false
        - name: email
          description: Pass to retrieve user by email
          in: query
          type: string
          required: false
        - name: offset
          in: query
          type: number
          required: false
        - name: limit
          in: query
          type: number
          required: false
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: '#/definitions/userShortType'
              count:
                type: number
              count_cached:
                type: boolean
              offset:
                type: number
              limit:
                type: number

    post:
      security:
      - accessCode: ['write_users']
      - implicit: ['write_users']
      - password: ['write_users']
      tags:
        - users
      description: Create a new user; only admin users can call this end point
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/userCreateType'
      responses:
        '204':
          description: Created
          schema:
            $ref: '#/definitions/userType'
        '403':
          description: Not allowed
  '/users/{userId}':
    get:
      security:
      - accessCode: ['read_users']
      - implicit: ['read_users']
      - password: ['read_users']
      tags:
      - users
      parameters:
        - name: userId
          in: path
          description: The ID of the user to retrieve
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/userType'
        '403':
          description: Not allowed
    patch:
      security:
      - accessCode: ['write_users']
      - implicit: ['write_users']
      - password: ['write_users']
      tags:
        - users
      parameters:
        - name: userId
          in: path
          description: The ID of the user to patch
          type: string
          required: true
        - name: user
          description: The user data to patch
          in: body
          schema:
            $ref: '#/definitions/userCreateType'
          required: true
        - name: X-VerificationId
          in: header
          description: 'A verification ID, this is used for updating lost passwords'
          type: string
          required: false
      responses:
        '200':
          description: Patched
          schema:
            $ref: '#/definitions/userType'
        '403':
          description: Not allowed
    delete:
      security:
      - accessCode: ['write_users']
      - implicit: ['write_users']
      - password: ['write_users']
      tags:
        - users
      parameters:
        - name: userId
          in: path
          description: The ID of the user to delete
          type: string
          required: true
      responses:
        '204':
          description: Deleted
        '403':
          description: Not allowed
...
  '/applications':
    get:
      description: Retrieve a list of all applications (admin)
      security:
      - accessCode: ['read_applications']
      - implicit: ['read_applications']
      - password: ['read_applications']
      tags:
        - applications
      parameters:
      - name: offset
        in: query
        type: number
        required: false
      - name: limit
        in: query
        type: number
        required: false
      - name: filter
        in: query
        type: string
        required: false
        description: URI-encoded JSON with filter terms (substring search), e.g. {"name":"michael"}. Only works with embed=1.
      - name: order_by
        in: query
        type: string
        required: false
        description: Ordering of the list, e.g. "name ASC", or "id DESC". Only works with embed=1.
      - name: embed
        in: query
        type: number
        required: false
        description: Pass embed=1 to embed the application data, otherwise only a list of IDs is passed on
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  title: AppId
                  type: object
                  properties:
                    id:
                      type: string
              count:
                type: number
              count_cached:
                type: boolean
              offset:
                type: number
              limit:
                type: number
        '403':
          description: Not allowed
    post:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      parameters:
        - name: application
          in: body
          schema:
            $ref: '#/definitions/applicationCreateType'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/applicationType'
...
  '/applications/{appId}':
    get:
      security:
      - accessCode: ['read_applications']
      - implicit: ['read_applications']
      - password: ['read_applications']
      tags:
        - applications
      parameters:
        - name: appId
          in: path
          description: ID of Application to retrieve
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/applicationType'
        '403':
          description: Not allowed
    patch:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      parameters:
        - name: appId
          in: path
          description: ID of Application to create or update
          type: string
          required: true
        - name: application
          in: body
          description: Application entity to patch
          schema:
            $ref: '#/definitions/applicationCreateType'
          required: true
      responses:
        '200':
          description: 'Success, patched'
          schema:
            $ref: '#/definitions/applicationType'
        '400':
          description: Bad request
        '403':
          description: Not allowed
    delete:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      parameters:
        - name: appId
          in: path
          description: ID of Application to delete
          type: string
          required: true
      responses:
        '204':
          description: 'Deleted, no content'
        '403':
          description: Not allowed
...
  '/auditlog':
    get:
      security:
      - accessCode: ['read_auditlog']
      - implicit: ['read_auditlog']
      - password: ['read_auditlog']
      tags:
        - misc
      description: Gets all audit logs (admin)
      responses:
        '200':
          description: Successful request
          schema:
            type: array
            items:
              $ref: '#/definitions/auditlogType'    
  '/approvals':
    get:
      security:
      - accessCode: ['read_approvals']
      - implicit: ['read_approvals']
      - password: ['read_approvals']
      tags:
        - misc
      description: Gets all pending approvals (admin)
      responses:
        '200':
          description: Successful request
          schema:
            type: array
            items:
              $ref: '#/definitions/approvalType'
  '/approvals/{approvalId}':
    get:
      security:
      - accessCode: ['read_approvals']
      - implicit: ['read_approvals']
      - password: ['read_approvals']
      tags:
        - misc
      description: Gets all pending approvals (admin)
      parameters:
        - name: approvalId
          in: path
          type: string
          required: true
          description: The ID of the approval record to retrieve
      responses:
        '200':
          description: Successful request
          schema:
            type: array
            items:
              $ref: '#/definitions/approvalType'
        '403':
          description: Not allowed
        '404':
          description: Not found
  '/verifications':
    post:
      security:
      - accessCode: ['write_verifications']
      - implicit: ['write_verifications']
      - password: ['write_verifications']
      tags:
        - verifications
      description: >-
        Registers for a new verification of email or a request for password
        reset
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/verificationType'
      responses:
        '204':
          description: >-
            No content. The API will not return whether it was successful or
            not.
        '400':
          description: Bad request
    get:
      security:
      - accessCode: ['read_verifications']
      - implicit: ['read_verifications']
      - password: ['read_verifications']
      tags:
        - verifications
      description: Returns a list of pending verifications (admin)
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/verificationType'
        '403':
          description: Not allowed.
  '/verifications/{verificationId}':
    get:
      security:
      - accessCode: ['read_verifications']
      - implicit: ['read_verifications']
      - password: ['read_verifications']
      tags:
        - verifications
      parameters:
        - name: verificationId
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/verificationType'
        '404':
          description: Not found
    delete:
      security:
      - accessCode: ['write_verifications']
      - implicit: ['write_verifications']
      - password: ['write_verifications']
      tags:
        - verifications
      parameters:
        - name: verificationId
          in: path
          type: string
          required: true
      responses:
        '204':
          description: 'Success, no content'
        '404':
          description: Not found.
...
  '/auth-servers':
    get:
      security:
      - accessCode: ['read_auth_servers']
      - implicit: ['read_auth_servers']
      - password: ['read_auth_servers']
      - clientCredentials: ['read_auth_servers']
      description: Get a list of available authorization servers
      tags:
        - misc
      responses:
        200:
          description: Successful response
          schema:
            type: array
            items:
              type: string
  '/auth-servers/{serverId}':
    get:
      security:
      - accessCode: ['read_auth_servers']
      - implicit: ['read_auth_servers']
      - password: ['read_auth_servers']
      - clientCredentials: ['read_auth_servers']
      description: Get information on a registered authorization serverId
      tags:
        - misc
      parameters:
        - name: serverId
          in: path
          required: true
          description: The ID of the authorization server to retrieve
          type: string
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/authServerType'
        404:
          description: Not found
  '/subscriptions/{client_id}':
    get:
      security:
      - accessCode: ['read_subscriptions']
      - implicit: ['read_subscriptions']
      - password: ['read_subscriptions']
      tags:
        - subscriptions
      description: For looking up a subscription by client_id (used for OAuth2 Implicit Grant)
      parameters:
        - name: client_id
          required: true
          description: The client_id created when creating a subscription
          in: path
          type: string
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              application:
                type: object
              subscription:
                type: object
        404:
          description: Not found.
  '/subscriptions':
    get:
      description: Retrieve a list of all subscriptions (admin/approvers)
      security:
      - accessCode: ['read_subscriptions']
      - implicit: ['read_subscriptions']
      - password: ['read_subscriptions']
      tags:
        - subscriptions
      parameters:
        - name: offset
          in: query
          type: number
          required: false
        - name: limit
          in: query
          type: number
          required: false
        - name: filter
          in: query
          type: string
          required: false
          description: URI-encoded JSON with filter terms (substring search), e.g. {"application_name":"demoapp"}. Only works with embed=1.
        - name: order_by
          in: query
          type: string
          required: false
          description: Ordering of the list, e.g. "application_name ASC", or "application_name DESC". Only works with embed=1.
        - name: embed
          in: query
          type: number
          required: false
          description: Pass embed=1 to embed the subscriptions data, otherwise only a list of IDs is passed on
      responses:
        200:
          description: Successful request
          schema:
            title: ArrayOfSubscriptions
            type: array
            items:
              $ref: '#/definitions/subscriptionType'
        403:
          description: Not allowed     
...
  '/pools':
    get:
      security:
      - accessCode: []
      - implicit: []
      - password: []
      - clientCredentials: []
      tags:
        - registrations
      description: Read registration pool information
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/poolMap'
  '/pools/{poolId}':
    get:
      security:
      - accessCode: []
      - implicit: []
      - password: []
      - clientCredentials: []
      tags:
        - registrations
      description: Read a single registration pool information
      parameters:
        - name: poolId
          in: path
          required: true
          type: string
          description: The ID of the pool to retrieve
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/poolType'
...
  '/grants/{userId}':
    get:
      security:
      - accessCode: ['read_grants']
      - implicit: ['read_grants']
      - password: ['read_grants']
      tags:
      - grants
      description: Retrieve a list of all user API grants
      parameters:
      - name: userId
        in: path
        type: string
        required: true
      - name: offset
        in: query
        type: number
        required: false
      - name: limit
        in: query
        type: number
        required: false
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/grantCollection'
        '400':
          description: Invalid Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      security:
      - accessCode: ['write_grants']
      - implicit: ['write_grants']
      - password: ['write_grants']
      tags:
      - grants
      description: Delete all of a user's API grants
      parameters:
      - name: userId
        in: path
        type: string
        required: true
      responses:
        '204':
          description: Success, no result
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/grants/{userId}/applications/{applicationId}/apis/{apiId}':
    parameters:
    - name: userId
      in: path
      type: string
      required: true
    - name: applicationId
      in: path
      type: string
      required: true
    - name: apiId
      in: path
      type: string
      required: true
      
    get:
      security:
      - accessCode: ['read_grants']
      - implicit: ['read_grants']
      - password: ['read_grants']
      tags:
      - grants
      description: Read a user's grants for a specific API
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/grantType'
        '403':
          description: Forbidden
        '404':
          description: Not found

    put:
      security:
      - accessCode: ['write_grants']
      - implicit: ['write_grants']
      - password: ['write_grants']
      tags:
      - grants
      description: Write a user API grant
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/grantType'
      responses:
        '204':
          description: Success, no result
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not Found

    delete:
      security:
      - accessCode: ['write_grants']
      - implicit: ['write_grants']
      - password: ['write_grants']
      tags:
      - grants
      description: Delete a specific user API grant
      responses:
        '204':
          description: Success, no result
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/accesstokens':
    post:
      security:
      - accessCode: ['access_tokens']
      - implicit: ['access_tokens']
      - password: ['access_tokens']
      tags:
      - accesstokens
      description: Add an access token registration
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/accessTokenType'
      responses:
        '201':
          description: Success, created
        '403':
          description: Forbidden
        '400':
          description: Bad request
    get:
      security:
      - accessCode: ['access_tokens']
      - implicit: ['access_tokens']
      - password: ['access_tokens']
      tags:
      - accesstokens
      description: Get access tokens by access token, refresh token or authenticated user id
      parameters:
      - name: access_token
        in: query
        required: false
        type: string
      - name: refresh_token
        in: query
        required: false
        type: string
      - name: authenticated_userid
        in: query
        required: false
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/accessTokenCollection'
        '403':
          description: Forbidden

    delete:
      security:
      - accessCode: ['access_tokens']
      - implicit: ['access_tokens']
      - password: ['access_tokens']
      tags:
      - accesstokens
      description: Delete access tokens by access token, refresh token or authenticated user id
      parameters:
      - name: access_token
        in: query
        required: false
        type: string
      - name: refresh_token
        in: query
        required: false
        type: string
      - name: authenticated_userid
        in: query
        required: false
        type: string
      responses:
        '204':
          description: Success
        '403':
          description: Forbidden
definitions:
  apiType:
    title: Api
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      desc:
        type: string
      plans:
        type: array
        items:
          type: string
      requiredGroup:
        description: 'Specify to restrict access to API, leave out to keep public'
        type: string
      tags:
        description: 'Array of tags for this API, use as "Category"'
        type: array
        items:
          type: string
      registrationPool:
        type: string
      passthroughUsers:
        type: boolean
      passthroughScopeUrl:
        type: string
      authMethods:
        description: Valid auth methods for this API, e.g. ["default:local", "default:google"]
        type: array
        items:
          type: string
...
  hrefType:
    type: object
    properties:
      href:
        type: string
...
  poolMap:
    type: object
    properties:
      any_property:
        $ref: '#/definitions/poolType'
...